treewide: implement serializer interface for socket API
For the socket API, the cmcd and testtool used CBOR
serialization for the payloads (regardless of the
serialization of metadata and attestation reports).

This commit introduces the socketApiSerializer configuration
for the testtool and adds automatic serialization detection
for socket API requests in the cmcd. This makes it possible
to interact with the cmcd socket API without requiring
CBOR support.

Signed-off-by: Simon Ott <simon.ott@aisec.fraunhofer.de>